+++ src/app/services/init/app-init.spec.ts
@@ -27,6 +27,7 @@ describe('appInit', () => {
       zodiacFontFamily: () => '',
       aspectImage: { width: 0, height: 0 },
       horoscoImage: { width: 0, height: 0 },
+      promittorImage: { width: 0, height: 0 },
       synastryAspectImage: { width: 0, height: 0 },
       synastryHoroscoImage: { width: 0, height: 0 },
       httpOptions: { headers: {} as any },